+++ R/MI.SANDRE.R
@@ -1,12 +1,13 @@
 #' Ajout des taxons et codes SANDRE
 #'
-#' Ajout des codes SANDRE et du nom du taxon de rang inférieur (espèce -> genre -> etc.) à partir d'un tableau issu de MI.systematique
+#' Ajout des codes SANDRE et du nom du taxon de rang inférieur (espèce -> genre -> etc.) à partir d'un tableau issu de \code{MI.systematique}
 #' @name MI.SANDRE
 #' @keywords donnees
 #' @import tidyverse
 #' @export
 #' @examples
 #' MI.SANDRE(data)
+#' data %>% MI.systematique() %>% MI.SANDRE()
 
 MI.SANDRE <- function(data)
 {